==> 0-ask_stable_beluga.ipynb <==
In this Script I build the "teacher" observations.

I will make use of an LLM, in this case `stabilityai/StableBeluga-7B` in order to get the labels that I will be using during the training phase of our lightweight BERT model.

import torch
from transformers import AutoModelForCausalLM, AutoTokenizer, pipeline
import json
from time import time
from sentence_transformers import SentenceTransformer
import nltk
import pandas as pd
import numpy as np
from sklearn.metrics.pairwise import cosine_similarity
from copy import deepcopy
import ast
# local tools
from local_tools import get_most_relevant_sentences, print_progress_bar, options_to_numerated_list

print("Available devices ", torch.cuda.device_count())
for i in range(torch.cuda.device_count()):
    print(f'Device {i}:', torch.cuda.get_device_name(i))

## Load Stable Beluge to the GPU

tokenizer = AutoTokenizer.from_pretrained(
    "stabilityai/StableBeluga-7B", 
    use_fast=True
)
model = AutoModelForCausalLM.from_pretrained(
    "stabilityai/StableBeluga-7B", 
    torch_dtype=torch.bfloat16,
    #low_cpu_mem_usage=True, 
    #device_map="auto"
)#.to("cuda")
model = model.to('cuda:0')

# We will be using this embedding model in order to shorten the news article when prompt is very long
embedding_model = SentenceTransformer("all-MiniLM-L6-v2").to('cuda:0')

SYS_PROMPT = "### System:\nYou are StableBeluga, an AI that follows instructions extremely well. Help as much as you can. Remember, be safe, and don't do anything illegal.\n\n"

# Query string that will be used to ask questions to Stable Beluga
Q = """
Given this news article:

"{news_text}"

When the article mentions "{entity}", which of the following options is the news article most likely referring to? Provide only one option.
options:

{search_options}
"""

# function to make the complete prompt that will be sent to Stable Beluga
make_prompt = lambda message: "{sys_prompt}### User: {message}\n\n### Assistant:\n The news article is most likely referring to: ".format(
    sys_prompt=SYS_PROMPT,
    message=message
)

def find_answer(sb_response):
    assis_token = "### Assistant:\n "
    end_of_answer = "</s>"
    soa_ix = sb_response.find(assis_token)
    eoa_ix = sb_response.find(end_of_answer)
    sb_answer = sb_response[soa_ix+len(assis_token):eoa_ix]
    return sb_answer

def make_tokenized_prompt(article_text, entity, search_options, top_n_perc=0.6):
    if top_n_perc < 0.1:
        return None
    question = Q.format(
        news_text = article_text,
        entity = entity,
        search_options=search_options
    )
    prompt=make_prompt(question)
    inputs = tokenizer(prompt, return_tensors="pt")["input_ids"]
    tokens_length = len(inputs[0])
    if tokens_length <= 4096:
        return inputs
    else:
        #print("token lengths were greater than 4096, summarizing the most relevant sentences")
        # summarize only the most importante sentences of the article
        most_relevant_pieces = get_most_relevant_sentences(
            article_text, 
            embedding_model, 
            top_n_perc = top_n_perc
        )
        return make_tokenized_prompt(
            article_text=most_relevant_pieces, 
            entity=entity, 
            search_options=search_options,
            top_n_perc=top_n_perc-0.2
        )

def ask_sb(input_tokens):
    if input_tokens is None:
        return ""
    inputs = input_tokens.to("cuda:0")
    with torch.no_grad():
        output = model.generate(
            inputs, 
            do_sample=False, 
            #top_p=0.95, 
            #top_k=0, 
            max_new_tokens=300
        )
    response = tokenizer.decode(output[0], skip_special_tokens=False)
    answer = find_answer(response)
    return answer

## Load DataSets
`ners_extracted_sofar.json` is a 2 GB file which will not be uploaded to github.
Contains 2,953,563 Named Entities for a subset of 112,522 news articles. 
I extracted Entities using Stanford Core NLP Python client. I focused only on PERSON, ORGANIZATION, COUNTRY and LOCATION entities.

For each entity of interesIwe performed a query string search to Wikidata knowledge graph, which sometimes returned several results for each entit.


Columns:

 - `text`: Is the text of the Named Entity. For example, Donald Trump.
 - `ner`: Is the type of the Named Entity, such as PERSON, LOCATION, etc.
 - `nerConfidences`: Confidence of the Named Entity belonging to a NER.
 - `clean_text`: Is the text of the Named Entity cleaned regarding special caracters and upper cased.
 - `h1`: Is the title of the news articles. Is used as key to merge with the news articles information.
 - `wikidata_search_entries`: Is a list of all instances of an entity that were found in Wikidata. Only one of those instances corresponds to the correct option.


`news_with_metadata.json` is a 1.5 GB file which will not be uploaded to github.
Contains 117,786 spanish news articles and metadata associated to those news articles, which was also extracted using other NLP techniques, such as translations and summarizations. Data was scrapped from "El Universal" digital news paper.

Relevant Columns:
 - `h1`: Is the title of the news articles. Is used as key to merge with the news articles information.
 - `date`: Date and time when the news articles was published.
 - `author`: Name of the author who published the article.
 - `content`: Article text in spanish.
 - `h1_en`: Article title translated to english.
 - `content_en`Article text translated to English.

path_file = 'datasets/ners_extracted_sofar.json'
with open(path_file, 'r') as jfile:
    ners_data = json.load(jfile)

ners_df = pd.DataFrame.from_dict(ners_data)

# free memory
try:
    # delete variable
    del ners_data
except NameError:
    # variable has already been deleted
    pass

ners_df.shape

# 'h1' if the title of the news article
len(ners_df['h1'].unique())

ners_df.head(20)

path_file = 'datasets/news_with_metadata.json'
with open(path_file, 'r') as jfile:
    processed_news_articles = json.load(jfile)

processed_news_articles_df = pd.DataFrame.from_dict(processed_news_articles)

processed_news_articles_df.shape

processed_news_articles_df.head(5)

# free memory
try:
    # delete variable
    del processed_news_articles
except NameError:
    # variable has already been deleted
    pass

# temporary file that will hold checkopoints in case that the process breaks.
path_file = 'ask_sb_sofar.json'
try:
    # open the checkpoint
    with open(path_file, 'r') as jfile:
        asked_sb_sofar = json.load(jfile)
    already_asked_sb = pd.DataFrame.from_dict(asked_sb_sofar)
    # index to identify where the process stopped, becomes entity + title
    already_asked_sb['ix'] = already_asked_sb['text']+'-'+already_asked_sb['h1']
except FileNotFoundError:
    # no checkpoint has been saved so far, process will start from the beginning
    asked_sb_sofar = []

ners_df['ix'] = ners_df['text']+'-'+ners_df['h1']

if len(asked_sb_sofar) > 0:
    to_ask_sb = ners_df[
        (ners_df['wikidata_search_entries'].apply(len) >1 )&
        ~(ners_df['ix'].isin(already_asked_sb['ix']))
    ]
else:
    to_ask_sb = ners_df[
        (ners_df['wikidata_search_entries'].apply(len) >1 )
    ]

to_ask_sb.head(5)

# free memory
try:
    # delete variable
    del ners_df
except NameError:
    # variable has already been deleted
    pass

# Example for one entry point in the process. We will use "Andres Manuel Lopez Obrador" as en example
example = to_ask_sb[to_ask_sb['text'] == 'Andres Manuel Lopez Obrador'].iloc[0]
print(example)

title = example['h1']
entity = example['text']
text = processed_news_articles_df[
    processed_news_articles_df['h1'] == title
]['content_en'].iloc[0]
options = options_to_numerated_list(
    [
        i.get("label_desc") for i in example['wikidata_search_entries']
    ]
)
sb_question_tokens = make_tokenized_prompt(
    article_text=text, 
    entity=entity, 
    search_options=options, 
    top_n_perc=0.6
)
question = Q.format(
    news_text = text,
    entity = entity,
    search_options=options
)
print("Stable Beluga Query: \n{0}".format(question))
sb_answer = ask_sb(sb_question_tokens)
print("Stable Beluga Disambiguation answer: \n{0}".format(sb_answer))

## Trigger the Process

Now we run the above process for the whole 2,953,563 Named Entities on a subset of 112,522 news articles on a loop. This way, Stable Belgua will help us create a big dataset of disambiguated observations which we can later use to fine tune a BERT model. The objective of fine tuning BERT is to have a lightweight model that can dismabiguate as good as Stable Belgua 7B.

The whole Stable Belgua disambiguation process took around 1 week on my 4090 GPU.

A lightweight model, as we will later see, uses 3x less GPU memory (which translates to lower GPU prouctive costs) and performs the disambiguation 3x faster (which also translates in faster productive systems).

sb_results = asked_sb_sofar
total_entries = len(to_ask_sb.to_dict(orient='records'))
for i, entry in enumerate(to_ask_sb.to_dict(orient='records')):
    if i%1000 == 0:
        torch.cuda.empty_cache()
        
    new_entry = deepcopy(entry)
    title = entry['h1']
    entity = entry['text']
    text = processed_news_articles_df[
        processed_news_articles_df['h1'] == title
    ]['content_en'].iloc[0]
    options = options_to_numerated_list(
        [
            i.get("label_desc") for i in entry['wikidata_search_entries']
        ]
    )
    sb_question_tokens = make_tokenized_prompt(
        article_text=text, 
        entity=entity, 
        search_options=options, 
        top_n_perc=0.6
    )
    sb_answer = ask_sb(sb_question_tokens)
    new_entry['sb_answer'] = sb_answer
    new_entry['options_given'] = options
    sb_results.append(new_entry)

    if i%1000 == 0:
        print_progress_bar(
            iteration=i, 
            total=total_entries, 
            bar_length=50
        )
        with open("ask_sb_sofar.json", "w") as file:
            json.dump(sb_results, file)

with open("datasets/sb_disambiguation_result.json", "w") as file:
    json.dump(sb_results, file)
